add clients for product and user services into cart; change service and repository implemntation in cart service

+++ cart-service/internal/services/cart.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/wazwki/WearStore/cart-service/internal/clients"
 	"github.com/wazwki/WearStore/cart-service/internal/domain"
 	"github.com/wazwki/WearStore/cart-service/internal/repository"
 )
@@ -13,30 +14,47 @@ type ServiceInterface interface {
 	RemoveFromCart(ctx context.Context, user_id, product_id string, quantity int) (bool, error)
 	GetCart(ctx context.Context, user_id string) (*domain.Cart, error)
 	ClearCart(ctx context.Context, user_id string) (bool, error)
-	Checkout(ctx context.Context, user_id string) (bool, float64, error)
+	Checkout(ctx context.Context, user_id string) (float64, error)
 }
 
 type Service struct {
-	repo repository.RepositoryInterface
+	repo    repository.RepositoryInterface
+	product clients.ProductClient
+	user    clients.UserClient
 }
 
-func NewService(repo repository.RepositoryInterface) ServiceInterface {
-	return &Service{repo: repo}
+func NewService(repo repository.RepositoryInterface, product clients.ProductClient, user clients.UserClient) ServiceInterface {
+	return &Service{repo: repo, product: product, user: user}
 }
 
 func (s *Service) AddToCart(ctx context.Context, user_id, product_id string, quantity int) (bool, error) {
-	_, err := s.repo.Add(ctx, user_id, product_id, quantity)
+	if quantity <= 0 {
+		return false, fmt.Errorf("quantity must be greater than zero")
+	}
+
+	product, err := s.product.GetProduct(ctx, product_id)
 	if err != nil {
 		return false, err
 	}
-	return true, nil
+
+	ok, err := s.repo.Add(ctx, user_id, product)
+	if err != nil {
+		return false, err
+	}
+
+	return ok, nil
 }
 
 func (s *Service) RemoveFromCart(ctx context.Context, user_id, product_id string, quantity int) (bool, error) {
+	if quantity <= 0 {
+		return false, fmt.Errorf("quantity must be greater than zero")
+	}
+
 	_, err := s.repo.Delete(ctx, user_id, product_id, quantity)
 	if err != nil {
 		return false, err
 	}
+
 	return true, nil
 }
 
@@ -45,6 +63,7 @@ func (s *Service) GetCart(ctx context.Context, user_id string) (*domain.Cart, er
 	if err != nil {
 		return nil, err
 	}
+
 	return cart, nil
 }
 
@@ -56,17 +75,46 @@ func (s *Service) ClearCart(ctx context.Context, user_id string) (bool, error) {
 	return success, nil
 }
 
-func (s *Service) Checkout(ctx context.Context, user_id string) (bool, float64, error) {
+func (s *Service) Checkout(ctx context.Context, user_id string) (float64, error) {
+	_, err := s.user.GetUser(ctx, user_id)
+	if err != nil {
+		return 0, err
+	}
+
 	cart, err := s.repo.Get(ctx, user_id)
 	if err != nil {
-		return false, 0, err
+		return 0, err
 	}
+
+	var totalCost float64
+	for _, item := range cart.Items {
+
+		product, err := s.product.GetProduct(ctx, item.ProductID)
+		if err != nil {
+			return 0, err
+		}
+
+		if product.Quantity < item.Quantity {
+			return 0, fmt.Errorf("not enough stock for product %v", item.ProductID)
+		}
+
+		_, err = s.product.UpdateProduct(ctx, &domain.CartItem{
+			ProductID: item.ProductID,
+			Name:      product.Name,
+			Price:     product.Price,
+			Quantity:  item.Quantity,
+		})
+		if err != nil {
+			return 0, err
+		}
+
+		totalCost += item.Price * float64(item.Quantity)
+	}
+
 	success, err := s.repo.Clear(ctx, user_id)
-	if err != nil {
-		return false, 0, err
+	if err != nil || !success {
+		return 0, fmt.Errorf("failed to clear cart after checkout: %w", err)
 	}
-	if !success {
-		return false, 0, fmt.Errorf("error clearing cart: %v", err)
-	}
-	return success, cart.TotalCost, nil
+
+	return totalCost, nil
 }
